oneToMany,oneToOne

+++ lab10b/studentmgmt/src/main/java/org/miu/cs/cs425/studentmgmt/service/StudentService.java
@@ -4,6 +4,6 @@ import org.miu.cs.cs425.studentmgmt.model.Student;
 
 public interface StudentService {
     void getAllStudents();
-    Student getStudentById(Integer id);
+    Student getStudentById(Long id);
     Student saveStudent(Student student);
 }
